+++ nekrosis/support/windows.py
@@ -5,7 +5,7 @@ windows.py: Windows-specific persistence logic.
 from .base import Persistence
 from .windows_utilities.startup import StartupFolder
 from .windows_utilities.persistence_methods import WindowsPersistenceMethods
-from .windows_utilities.shortcut import TaskbarShortcut
+
 
 class WindowsPersistence(Persistence):
     """
@@ -56,14 +56,13 @@ class WindowsPersistence(Persistence):
         """
         method = self.configured_persistence_method()
 
+        
+
         if method == WindowsPersistenceMethods.STARTUP_CURRENT_USER.value:
             StartupFolder(payload=self.payload).install_current_user()
             return
         if method == WindowsPersistenceMethods.STARTUP_GLOBAL.value:
             StartupFolder(payload=self.payload).install_global()
             return
-        if method == WindowsPersistenceMethods.SHORTCUT_USER.value:
-            TaskbarShortcut(payload=self.payload).install()
-            return
-            
+
         raise NotImplementedError(f"Method {method} not implemented.")